+++ NEWS.md
@@ -1,3 +1,13 @@
+# ggpubr 0.2.5.999
+  
+  
+## New features
+
+
+## Minor changes
+
+## Bug fixes
+
 # ggpubr 0.2.5
 
 ## Minor changes